==> 06-interactive.Rmd <==
# Interactive component


```{r}
library(tidyverse)
library(patchwork)
library(ggnewscale)
library(rtweet)
library(sjmisc)
library(readr)
library(base)
```

```{r}
df <- read_csv("./resources/data_combined_filtered_sentiment.csv")
```

```{r}
df_t <- df %>% select(created_at, group, sentiment_score)
```

```{r}
times <- as.character(df_t$created_at)

dates <- c()

for (time in times){
  date <- strsplit(time, " +")[[1]][1]
  dates <- c(dates, date)
}

df_t$created_at <- dates
```

```{r}
df_p <- df_t %>% filter(group == "politicians") 
df_m <- df_t %>% filter(group == "medical")
df_c <- df_t %>% filter(group == "celebrities")

df_p <- df_p %>%
  group_by(created_at) %>%
  summarise(sentiment_score = mean(sentiment_score))

df_m <- df_m %>%
  group_by(created_at) %>%
  summarise(sentiment_score = mean(sentiment_score))

df_c <- df_c %>%
  group_by(created_at) %>%
  summarise(sentiment_score = mean(sentiment_score))

df_combined <- merge(df_p, df_m, by = "created_at")
df_combined <- merge(df_combined, df_c, by = "created_at")
colnames(df_combined) <- c('x', 'politicians', 'medicals', 'celebrities')

df_combined_na <- merge(df_p, df_m, by = "created_at", all = TRUE)
df_combined_na <- merge(df_combined_na, df_c, by = "created_at", all = TRUE)
colnames(df_combined_na) <- c('x', 'politicians', 'medicals', 'celebrities')

write_csv(df_combined_na, "./resources/interactive_with_na.csv")
write_csv(df_combined, "./resources/interactive_without_na.csv")
```

```{r, echo=FALSE, fig.width=8}
htmltools::includeHTML("interactive.html")
```


On this chapter, we made an interactive graph where you can see the difference of sentiment score transitions among different groups. Our goal of this project is to see how different groups react against the same news topic, here COVID-19. By making this graph interactive, the graph dynamically illustrates the difference between two sentiment lines, which enables us to observe the difference more intuitively than the static graph in the section 5.2.1. Using the drop-down button at top center, we can choose which group's sentiment score graph over time to display. We smoothed the curve by using d3 library d3-regression because we observed many dates when no user tweeted about COVID-19. The graph will start by showing politicians' sentiment score graph by default, and choosing another group from the button will transform the graph accordingly. The benefit of using this interactive graph is that we can see each group's sentiment score changes more closely, and also, we can tell the difference between two groups by seeing how the graph transforms as we choose another group from the button.

*Note: We made two csv files for interactive graph. Both csv contain four columns - date, politicians' sentiment scores, medicals' sentiment scores, celebrities' sentiment scores. interactive_with_na.csv contains all dates that at least one of the groups has sentiment scores, which means there are NAs in some entries for each group. interactive_without_na.csv only contains dates that all three groups have sentiment scores. We decided to use interactive_without_na.csv because having NAs in the dataset made the interactive graph malfunction when transitioning to another group.*

As we can see from the graph, sentiment score of politicians move in the opposite direction of the medical's score. One can go back and forth between the two groups' graphs to see that the general flow of the graphs change dramatically. Though the sentiment score of celebrities generally fluctuated more than the other groups, the sentiment scores are generally higher than the others' after COVID-19. In this way, we could observe that there are differences of reactions among different groups from this interactive graph.
