==> 07-conclusion.Rmd <==
# Conclusion

We set out this project to observe how similar or different people (celebrities, politicians, and medical) react to the same news (COVID-19). We used Twitter API to gather our data and conducted sentiment analysis in R as a way to measure people's reaction. Our analysis was over the Tweets posted from March 2020 to November 2021. 


Some of the limitations of our projects are as follows: 

1) We could not directly pull the most followed accounts using API but used an internet source to identify the most followed accounts for each group.  

2) We used the tidytext package to perform the sentiment analysis for this project, but more advanced models involving deep learning methods can be implemented to increase the accuracy of the analysis.


We also identified opportunities to improve and expand our projects:

1) We can expand the sentiment analysis to different topics to see how sentimental reactions differ by group.

2) We can further explore the relationship between sentiments and retweets/likes to understand whether certain sentiments attract more likes. 

3) We can expand on the deep dive section (5.3) to further divide up a group into more granular subgroups, such as gender, age, or region.


Through our analysis and visualizations we learned useful insights. The key takeaways are as follows: 

1) Celebrities' reactions were generally more positive than the other two groups'.

2) The most dominant sentiment among medical group was "fear", and their sentiments did not change much over time. 

3) The most dominant sentiment among politicians was "trust", and their sentiments did not change significantly over time.
